% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{add_fun_opt}
\alias{add_fun_opt}
\title{Add functions to the options.}
\usage{
add_fun_opt(fun = "fun")
}
\arguments{
\item{fun}{character. name of the function}
}
\value{
nix
}
\description{
The added function is also tested for functionality.
}
\examples{

# init_stp25_options()

n_mean_sd <- function(x, digits, n,  ...) {
 if( is.na(digits)  ) digits <- 2
 data.frame(
  lev = c("   n", "   mean", "   sd"),
  n = c(as.character(n), "", ""),
  m = c(
    length(x),
    render_f(mean(x, na.rm = TRUE), digits),
    render_f(sd(x, na.rm = TRUE), digits)
  )
 )
}
 

add_fun_opt("n_mean_sd")
 
 get_opt("measure_fun")
 
# DF |>
#  Tbll_desc(
#    glucose,
#    harnstoff[n_mean_sd],
#    by = ~group
#  )
}
