% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice-auto-plot.R
\name{auto_plot}
\alias{auto_plot}
\alias{auto_plot.lm}
\alias{auto_plot.default}
\title{Lattice-Matrix-Plot}
\usage{
auto_plot(...)

\method{auto_plot}{lm}(x, ...)

\method{auto_plot}{default}(
  ...,
  origin = 0,
  xlab = NULL,
  ylab = NULL,
  type = c("p", "r"),
  cex.main = 1,
  cex.scales = 0.75,
  ncol = NULL,
  default.scales = list(abbreviate = TRUE, minlength = 5, cex = cex.scales),
  layout = NULL,
  lattice.options = list(layout.heights = list(axis.xlab.padding = list(x = 0),
    xlab.key.padding = list(x = 0))),
  col = farbe(),
  col.bar = "gray50",
  par.settings = bw_theme(col = col, col.bar = col.bar),
  include.n = TRUE,
  par.strip.text = NULL,
  wrap.main = NULL,
  grid.arrange = TRUE,
  levels.logical = c(TRUE, FALSE),
  labels.logical = levels.logical,
  include.label = TRUE,
  include.percent = FALSE,
  include.reorder = FALSE,
  include.reorder.last = NULL,
  main = ""
)
}
\arguments{
\item{...}{Variablen und Daten}

\item{ylab, xlab}{default ist ""}

\item{type}{c("histogram", "boxplot")}

\item{cex.main, cex.scales}{Ueberschrift und Scales}

\item{ncol}{an grid.arrange}

\item{par.settings}{sefault ist  par.settings = bw_theme((farbe()))}

\item{include.n}{noch nicht implemeniert}

\item{grid.arrange}{logical Arrange multiple grobs on a page}
}
\value{
lattice Plot
}
\description{
Lattice-Matrix-Plot
}
\examples{

enviro <- lattice::environmental

enviro <- transform(
  enviro,
  smell = cut(
    enviro$ozone,
    breaks = c(0, 30, 50, Inf),
    labels = c("ok", "hmmm", "yuck"),
    ordered = TRUE
  ),
  is.windy = factor(
    enviro$wind > 10,
    levels = c(TRUE, FALSE),
    labels = c("windy", "calm")
  )
)
#head(enviro)

auto_plot(enviro, ozone, radiation, is.windy, wind, smell, by =  ~ temperature)

auto_plot(ozone ~ radiation + is.windy + wind + smell, enviro)

auto_plot(enviro, ozone[box], radiation[hist], is.windy, wind, smell, temperature)

auto_plot(enviro, ozone, radiation, is.windy, wind, by =  ~ smell)

# das ist eine Liste mit plots
p1 <-
  auto_plot(enviro,
            ozone,
            radiation,
            is.windy,
            wind,
            smell,
            temperature,
            grid.arrange = FALSE)


for (i in seq_along(p1)) {
  cat("\n", names(p1[i]))
  print(p1[[i]])
  #  SavePlot(  names(p1[i]) , w=3.6, h=2.9)
}


}
