% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clensing-encoding.R
\name{clean_character}
\alias{clean_character}
\title{Clean and standardize character strings}
\usage{
clean_character(
  x,
  tolower = TRUE,
  abbreviate = FALSE,
  minlength = 12,
  ascii = FALSE,
  replace = c(`\%` = "_pct", `#` = "_cnt", `¨` = "", `&+` = "_and_", `@+` = "_at_", ä =
    "ae", ü = "ue", ö = "oe", Ü = "Ue", Ä = "Ae", Ö = "Oe", ß = "ss"),
  unique = TRUE
)
}
\arguments{
\item{x}{Character vector to be cleaned.}

\item{tolower}{Logical, convert to lower case? (default: TRUE)}

\item{abbreviate}{Logical, abbreviate names to minlength? (default: FALSE)}

\item{minlength}{Minimum length for abbreviations. (default: 12)}

\item{ascii}{Logical, Transliterates characters to ASCII}

\item{replace}{Named character vector of replacements.}

\item{unique}{Logical, make names unique? (default: TRUE)}
}
\value{
A cleaned character vector.
}
\description{
This function converts character strings into standardized,
machine-friendly names (e.g., for variable or file names).
}
\details{
Steps:
- Transliterates characters to ASCII
- Converts to lower case (optional)
- Applies user-defined replacements
- Removes unwanted characters
- Collapses multiple separators
- Optionally abbreviates long names
- Ensures uniqueness (optional)
}
\examples{
clean_character(c("Ölß", "Extra lang mit Müll!!!", "#1", "äüö_ ga"))

}
