% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-trans-long.R
\name{clean_names}
\alias{clean_names}
\title{Clean variable names}
\usage{
clean_names(vars, data)
}
\arguments{
\item{vars}{Character vector of variable names/positions}

\item{data}{The input data frame}
}
\value{
Cleaned variable names
}
\description{
Clean variable names
}
