% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\alias{save_codebook}
\alias{use_codebook}
\title{Codebook}
\usage{
save_codebook(x, file = "codebook.xlsx", include.data = FALSE)

codebook(x)

use_codebook(
  data = NULL,
  data_codebook = NULL,
  file = "demo.xlsx",
  sheet.data = "data",
  sheet.codebook = "codebook",
  names = "names",
  label = "label",
  value.labels = "value.labels"
)
}
\arguments{
\item{x}{Data.frame}

\item{file}{File (xlsx)}

\item{data}{data.frame, default is NULL}

\item{sheet.data, sheet.codebook}{file= "demo.xlsx", sheet.data = 1, sheet.codebook = 2,}

\item{names, label, value.labels}{Variablen NNamen im codebook}

\item{codebook}{data.frame codebook = codebook(DF), default is NULL}
}
\value{
data.frame

data.frame or tibble
}
\description{
Speichern des Codebook

codebook: Create a structure of the data as data.frame.

use_codebook: Restore the structure of the data from the 
Excel file created by the Codebook function.
}
\examples{
 \dontrun{

DF <- data.frame(
  BMI = c(1, 2, 3, 1, 2, 3),
  WHtR = gl(2, 3, label = c("Amy", "Bob")),
  WHtR_1 = c(9, 7, 6, 8, 6, 9),
  bildprof = c(6, 7, 8, 5, 6, 7)
  ) |>
  Label(
        BMI = "Body-Mass-Index",
        WHtR =  "Waist-Height-Ratio",
        WHtR_1 = "Waist-Height-Ratio"
        )
 
  save_codebook(DF, "codebook.xlsx" )
 
#' Labels und Factoren ubernehmen
DF2 <-   use_codebook(
  DF,
  file = "codebook.xlsx",
  sheet.codebook = 2
)

#' Nur die Labels  ubernehmen
DF3 <-   use_codebook(
  DF,
  file = "codebook.xlsx",
  sheet.codebook = 2,
  value.labels = NULL
)

#' Strucktur aus einem anderen data.frame uebernehmen
DF1 <- use_codebook(DF, DF2)

#' Daten und Strucktur im Excell-File
DF4 <-  use_codebook(file = "codebook.xlsx")
#readxl::read_excel("codebook.xlsx", sheet = "codebook")
#'
}
 
}
