% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{create_formulas}
\alias{create_formulas}
\title{Create Analysis Formulas}
\usage{
create_formulas(measure_vars, measure_vars_clean, by = NULL, groups = NULL)
}
\arguments{
\item{measure_vars}{Original measure variables}

\item{measure_vars_clean}{Cleaned measure variable names}

\item{by}{Grouping specification}
}
\value{
List with two components:
\itemize{
  \item formula_in: Formula with original variable specifications
  \item formula: Formula with cleaned variable names
}
}
\description{
Internal function to create both original and cleaned formulas.
}
\examples{
# example code
  stp25tools2:::create_formulas(
measure_vars = c("log(m1)", "m2", "m3", "m4"),
measure_vars_clean = c("log(m1)", "m2", "m3", "m4"),
by = ~ sex
)
stp25tools2:::create_formulas(
  measure_vars = c("log(m1)", "m2", "m3", "m4"),
  measure_vars_clean = c("log(m1)", "m2", "m3", "m4"),
  by = NULL
)
}
\keyword{internal}
