% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Cs}
\alias{Cs}
\alias{XLS}
\alias{paste_names}
\alias{paste_names_levels}
\title{Character strings from unquoted names}
\usage{
Cs(...)

XLS(...)

paste_names(x, collapse = ", ", ...)

paste_names_levels(x, abbreviate = TRUE, collapse = ", ", ...)
}
\arguments{
\item{...}{liste mit den Spaltennamen A:BB}

\item{x}{dataframe ofe factor}

\item{collapse}{an  character string to separate the results ", "}

\item{abbreviate}{logical.  in paste_names_levels}
}
\value{
Vektor

character

character
}
\description{
Erweiterung der  Hmisc Cs-Funktion um trim.ws.
Erlaubt Variablen mit Komma oder Plus(+) oder Leerzeichen abzuternnen.

XLS:

Nummernindex aus Excel - Spaltenbezeichnung

Extrahiert aus Buchstaben die Spaltennummer

paste_names:  paste names

paste_names_levels:
Copy and Paste.
}
\examples{

Cs(sd, fr, fg)
Cs(sd, fr, fg, "hju nh")
Cs("  Hallo Welt ")
Cs(~ sd + fr + fg)
as.roman(1968)
#strsplit("A:V", "\\:")
XLS(a, B)
XLS(a, B, c:f, g:h,i, r:z)
XLS(A:Z)

}
