% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-trans-apply.R
\name{Dapply}
\alias{Dapply}
\alias{Dapply.formula}
\alias{Dapply.data.frame}
\alias{dapply2}
\alias{fix_sosci}
\title{Apply Functions Over Data Frames}
\usage{
Dapply(...)

\method{Dapply}{formula}(x, data, fun = function(y) as.numeric(y), ...)

\method{Dapply}{data.frame}(data, ..., fun = function(y) as.numeric(y))

dapply2(data, fun = function(x) as.numeric(x), ...)

fix_sosci(
  data,
  measure_multi = "ausgewählt",
  na_strings = "nicht beantwortet",
  trim_whitespace = TRUE,
  recode_numeric_na = FALSE,
  na_values = c(-99, -98, -97, 999),
  ...
)
}
\arguments{
\item{...}{Namen der Spalten}

\item{x}{Objekt data.frame, formula}

\item{data}{Data.frame}

\item{fun}{funktion function(x) as.numeric(x)}

\item{measure_multi}{Multi-Response Items "ausgewählt"}

\item{na_strings}{NA-Strings aus Faktor entfernen "nicht beantwortet"}

\item{trim_whitespace}{Leehrzeichen bei character TRUE}

\item{recode_numeric_na, na_values}{Numerische Variablen behandeln FALSE and  c(-99, -98, -97, 999),}
}
\value{
data.frame
}
\description{
Dapply, dapply2: plyr::llply() + Label()

dapply2: Copie of plyr::llply()

Analog wie dapply2 aber speziell für SoSci-Umfragen
}
\examples{


df1 <- tibble::tibble(
  month = rep(1:3, 2),
  student = rep(c("Amy", "Bob"), each = 3),
  A = c(9, 7, 6, 8, 6, 9),
  B = c(6, 7, 8, 5, 6, 7),
  C = c(1.6, 2.7, 3.8, 4.5, 5.6, 6.7)
) |>
  Label(
    month = "Monat",
    student = "Schüler",
    A = "Deutsch",
    B = "Mathe"
  )






Dapply( ~ A + B, df1, ~ .x * 100)

Dapply(~ A + B, df1, fun = function(x) x * 100)

df1 |> Dapply( ~ A + B, fun = function(x) x * 100)

df1 |> Dapply(A, B, fun = function(x) x * 100)

df1 |> Dapply(A:C, fun = function(x) x * 100)

dapply2(df1, ~ as.numeric(factor(.x)))


#' Alternative mit across::across


dplyr::mutate(df1, dplyr::across(c("A", "B"), .fns= function(x) x * 100 ))


}
