% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{delet_label}
\alias{delet_label}
\title{Remove Variable Labels}
\usage{
delet_label(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
The data frame with all variable labels removed
}
\description{
Remove Variable Labels
}
