% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_likert.R
\name{doi_likert}
\alias{doi_likert}
\title{Dominant Opinion Index (DOI) für Likert-Skalen}
\usage{
doi_likert(positive, negative, neutral = NULL)
}
\arguments{
\item{positive}{Prozentualer Anteil positiver Antworten (z. B. Agree + Strongly agree).}

\item{negative}{Prozentualer Anteil negativer Antworten (z. B. Disagree + Strongly disagree).}

\item{neutral}{Prozentualer Anteil neutraler Antworten (Default = 100 - positive - negative).}
}
\value{
Numerischer Wert zwischen -100 und +100.
}
\description{
Dominant Opinion Index (DOI) für Likert-Skalen
}
\examples{
stp25tools2:::doi_likert(positive = 40, negative = 30, neutral = 30)

}
