% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_forest_epi.R
\name{example_compatible_facets}
\alias{example_compatible_facets}
\title{Beispiel für kompatible Facet-Plots}
\usage{
example_compatible_facets()
}
\value{
Liste mit Beispiel-Plots
}
\description{
Beispiel für kompatible Facet-Plots
}
\examples{
\dontrun{
example_compatible_facets()
}
}
