% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_multiple_choice.R
\name{extract_multiple_responses}
\alias{extract_multiple_responses}
\title{Extract Multiple Responses from Data Frame}
\usage{
extract_multiple_responses(
  data,
  sep = ",",
  rm_string = NULL,
  as_logical = TRUE
)
}
\arguments{
\item{data}{A data frame containing the multiple response columns}

\item{sep}{Character to split responses (default: ",")}

\item{rm_string}{Suchstring zu ersaetzen mit dem Seperator}

\item{as_logical}{output logical or integer (0,1)}
}
\value{
A data frame with binary variables for each unique response
}
\description{
Processes columns with multiple responses using commas and parentheses as separators
and creates binary variables for each unique response.
}
\examples{

dat <-  data.frame(
Q1 = c(12345, 3, 35, 134, 12,
       13,  14, 12, 123, 123,
       13,  14, 12, 123,123),
Q2 = c(
  "Alm Dudler, Mineral, Cola, Bier, Wein",
  " Cola, ",
  "Cola,  Wein",
  "Alm Dudler,  Cola, Bier",
  "Alm Dudler, Mineral",

  "Alm Dudler,  Cola",
  "Alm Dudler,  Bier,",
  "Alm Dudler, Mineral",
  "Alm Dudler, Mineral, Cola",
  "Alm Dudler, Mineral, Cola",

  "Alm Dudler, Cola",
  "Alm Dudler,  Bier",
  "Alm Dudler, Mineral, ",
  "Alm Dudler, Mineral, Cola",
  "Alm Dudler, Mineral, Cola"
)
)

extract_multiple_responses(dat[1])
extract_multiple_responses(dat[2] )


# library(UpSetR)
# library(ggplot2)

# Daten vorbereiten für UpSetR
# DF <- extract_multiple_responses(dat[2] )
#upset_data <- as.data.frame(lapply(DF, as.integer))
#upset(upset_data, nsets = ncol(upset_data), order.by = "freq")
}
