% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_classes}
\alias{get_classes}
\title{Get simplified variable classes}
\usage{
get_classes(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Get simplified variable classes
}
\keyword{internal}
