% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-trans-long.R
\name{get_measure_vars}
\alias{get_measure_vars}
\title{Get measure variables from dots}
\usage{
get_measure_vars(dots, data)
}
\arguments{
\item{dots}{Arguments (enquos)}

\item{data}{The input data frame}
}
\value{
Character vector of measure variables
}
\description{
Get measure variables from dots
}
