% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_opt}
\alias{get_opt}
\title{Get package options}
\usage{
get_opt(name = NULL, type = NULL)
}
\arguments{
\item{name}{(Optional) Name of the option group to retrieve}

\item{type}{(Optional) Specific parameter within the group to retrieve}
}
\value{
Requested options or values
}
\description{
Retrieve current package options or specific settings.
}
\examples{
get_opt()
get_opt("mean")
get_opt("mean", "digits")
}
