% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{init_stp25_options}
\alias{init_stp25_options}
\title{Initialize stp25 options from config file
config_path Path to config.yml file}
\usage{
init_stp25_options()
}
\description{
Initialize stp25 options from config file
config_path Path to config.yml file
}
