% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{Label}
\alias{Label}
\alias{delet_label}
\alias{set_label}
\alias{get_label}
\alias{trimm_label}
\alias{gsub_label}
\title{Label}
\usage{
Label(data, ...)

delet_label(data)

set_label(data, label = NULL)

get_label(data, ...)

trimm_label(data, pattern = "\\\\.\\\\.\\\\. ", pos = 2)

gsub_label(data, pattern = "\\\\&amp;", replacement = "&")
}
\arguments{
\item{data}{data.frame}

\item{...}{label in der Form a="Hallo, b="Welt"}

\item{label}{attribut label}

\item{pattern, replacement}{Pattern to look for.}

\item{pos}{an stringr::str_split_fixed}
}
\description{
Setzen des attr(data, "label")

Remove Variable Labels - returns the data frame with all variable labels removed

set_label

Get variable labels Returns A named vector of labels.

trimm_label, gsub_label
}
\examples{

DF <- data.frame(
  x = 1:6,
  y = 1:6,
  z = 1:6,
  BMI = c(1, 2, 3, 1, 2, 3),
  WHtR = gl(2, 3, labels = c("Amy", "Bob")),
  WHtR_1 = c(9, 7, 6, 8, 6, 9),
  bildprof = c(6, 7, 8, 5, 6, 7)
) |>
  Label(
    BMI = "Body-Mass-Index",
    WHtR =  "Waist-Height-Ratio",
    WHtR_1 = "Waist-Height-Ratio",
    dummy = "Dummy",
    dummy2 = "Dummy"
  )



get_label(DF)
get_label(DF)
DF <- set_label(DF, c(bildprof = "Bildungsprofil"))
get_label(DF)
DF <- delet_label(DF)
get_label(DF)

}
