% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice-auto-plot.R
\name{multi_barplot}
\alias{multi_barplot}
\title{multi_barplot}
\usage{
multi_barplot(
  ...,
  reorder = TRUE,
  last = NULL,
  main = NULL,
  ylab = "",
  include.percent = FALSE,
  origin = 0,
  xlab = if (include.percent) "Percent" else "Count",
  xlim = NULL,
  ylim = NULL,
  wrap = TRUE,
  use.level = 1
)
}
\arguments{
\item{...}{an Summarise}

\item{reorder, last}{an reorder2}

\item{main, ylab, origin, xlab}{an Lattice}

\item{include.percent}{summary}
}
\value{
lattice plot
}
\description{
Adaption von `lattice::barchart()` und Feineinstellung können
mit `lattice::update.trellis` erstellt werden.
}
\examples{

n <-99
DF <- data.frame(
  Magazines = rbinom(n, 1,prob=.75),
  Comic.books =rbinom(n, 1,prob=.25),
  Fiction = rbinom(n, 1,prob=.5),
  Sonstiges = rbinom(n, 1,prob=.35)
) |> transform(sex = cut(rnorm(n), 2, c("m", "f")))


multi_barplot(  DF, .~ sex, last="Sonstiges")


}
