% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll-xtabs-2x2.R
\name{praevalenz}
\alias{praevalenz}
\title{Numerische Berechnung von Prävalenzen mit Konfidenzintervallen}
\usage{
praevalenz(
  data,
  x,
  clopper.pearson = TRUE,
  wilson.score = FALSE,
  agresti.coull = FALSE,
  wald = FALSE,
  conf.level = 0.95,
  digits = 1L,
  referenz = NA
)
}
\description{
Berechnet Prävalenzen mit verschiedenen Methoden für Konfidenzintervalle
und gibt die Ergebnisse als numerische Werte für die weitere Verarbeitung
(z.B. Visualisierung) zurück.
}
