% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{prepare_output}
\alias{prepare_output}
\title{prepare_output}
\usage{
prepare_output(
  x,
  caption = "",
  note = "",
  N = NULL,
  labels = NA,
  rgroup = NULL,
  n.rgroup = NULL,
  ...
)
}
\arguments{
\item{x}{data.frame}

\item{caption}{character}

\item{note}{character}

\item{N}{number}

\item{labels}{optional}

\item{rgroup}{optional}

\item{n.rgroup}{optional}

\item{...}{not used}
}
\value{
tibble, data.frame
}
\description{
prepare_output
}
