% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rbind.R
\name{Rbind2}
\alias{Rbind2}
\title{Bind multiple data.frames with additional metadata}
\usage{
Rbind2(
  ...,
  .id = "which",
  .names = NULL,
  .use.label = TRUE,
  include.rownames = FALSE
)
}
\arguments{
\item{...}{Data frames oder tibbles, die gebunden werden sollen.}

\item{.id}{Name der Spalte, die die Herkunft kennzeichnet (Standard `"which"`).
Wenn `NULL`, wird keine Herkunftsspalte hinzugefügt.}

\item{.names}{Optionaler Vektor mit Namen für die Datenquellen.
Wenn `NULL`, werden die Namen aus den übergebenen Objekten ermittelt.}

\item{.use.label}{`logical`. Wenn `TRUE` (Default), werden Labels
von den Eingabe-Dataframes übernommen.}

\item{include.rownames}{`logical`. Falls `TRUE`, wird eine zusätzliche
Spalte `"Source"` mit den ursprünglichen Rownames ergänzt.}
}
\value{
Ein `data.frame` (genauer: `tibble`), das die Daten kombiniert.
}
\description{
`Rbind2()` ist ein Wrapper um [dplyr::bind_rows()], der mehrere
`data.frame`s zu einem gemeinsamen Datensatz kombiniert und dabei
zusätzliche Features bietet:
}
\details{
* fügt eine Spalte mit der Herkunft (`.id`) hinzu,
* erlaubt optional die Beibehaltung von Rownames,
* übernimmt Spaltenlabels (via `get_label()` und `set_label()`).
}
\examples{
df1 <- data.frame(a = 1, b = 2)
df2 <- data.frame(a = 2, c = 3, d = 5)

Rbind2(df1, df2)
#>   which a  b  c  d
#> 1   df1 1  2 NA NA
#> 2   df2 2 NA  3  5

df1 <- data.frame(CustomerId = 1:3, Product = "Oven")
df2 <- data.frame(CustomerId = 4:5, Product = "Television")
df3 <- data.frame(CustomerId = 6:7, Product = "Air conditioner", State = "NJ")

Rbind2(df1, df2, df3)

}
\seealso{
[dplyr::bind_rows()], [plyr::rbind.fill()]
}
