% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{read_text_table}
\alias{read_text_table}
\title{Read Data Frame from Text String}
\usage{
read_text_table(
  text_input,
  na.strings = c("NA", "na"),
  sep = "",
  dec = ".",
  strings_as_factors = TRUE
)
}
\arguments{
\item{text_input}{Character string containing the table data.}

\item{na.strings}{Character vector of strings to interpret as NA.}

\item{sep}{Field separator character (default: whitespace).}

\item{dec}{Decimal point character.}

\item{strings_as_factors}{Logical; convert character vectors to factors.}
}
\value{
A data frame.
}
\description{
Reads a data frame from a character string containing table-like text.
}
\examples{
read_text_table("
a b c
1 2 3
4 5 6
7 8 9
")

#   a b c
# 1 1 2 3
# 2 4 5 6
# 3 7 8 9
}
