% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{contr.Treatment}
\alias{contr.Sum}
\alias{contr.Helmert}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{car}{\code{\link[car:contr.Helmert]{contr.Helmert()}}, \code{\link[car:contr.Sum]{contr.Sum()}}, \code{\link[car:contr.Treatment]{contr.Treatment()}}}
}}

