% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-render.R
\name{render_f}
\alias{render_f}
\alias{render_sigf}
\alias{render_f.list}
\alias{render_f.matrix}
\alias{render_f.data.frame}
\alias{render_f.tbl_df}
\alias{render_f.numeric}
\alias{render_f.default}
\title{Format a vector of provided numeric values}
\usage{
render_f(x, digits = NULL, ...)

render_sigf(x, digits = 3)

\method{render_f}{list}(x, digits = NULL, ...)

\method{render_f}{matrix}(x, digits = NULL, ...)

\method{render_f}{data.frame}(x, digits = NULL, ...)

\method{render_f}{tbl_df}(x, digits = NULL, ...)

\method{render_f}{numeric}(x, digits = NULL, ...)

\method{render_f}{default}(
  x,
  digits = NULL,
  drop0leading = FALSE,
  na.strings = "NA",
  na.symbol = "",
  decimal.mark = OutDec(),
  ...
)
}
\arguments{
\item{x}{numeric; the data to format. Must work with quantile function.}

\item{digits}{.numeric}

\item{...}{render_f}

\item{drop0leading}{leading zero}

\item{na.strings, na.symbol}{NA replace}

\item{decimal.mark}{an formatC default format="f"}
}
\value{
character; formatted values as character strings
}
\description{
Given a vector of data return as strings formatted as requested

render_sigf:  Format numbers to significant figures nicely
}
\details{
Stolen from tangram
}
\examples{

 \donttest{
 # Vector
x <- c(1, 0, 2.45896, 2548.256)
y <- c(0.1, 0.01 , 0.001, NA)
z <- c(1L, 2L, 3L)
render_f(x)
render_f(y)
render_f(z)
render_f(as.double(x))
render_f(letters[1:3])

render_f(x, c(1, 2, 3, 4))
render_f(x, c(1, 2, 3))

# Data.Frame
dat <- data.frame(Item = letters[1:4], x = x, y = y)
render_f(dat)
render_f(dat, digits = 2)
render_f(dat, digits = c(1, 2))
render_f(dat, digits = c(1, 2, 3, 4, 1))
render_f(dat,
         digits = list(c(1, 2), c(3)),
         drop0leading = list(FALSE, TRUE))

# Matrix
render_f(cbind(y, x))
render_f(cbind(y, x) , digits = list(3, 0))
# List
mx <- list(x = x, y = y, z = rnorm(10))
render_f(mx)
render_f(mx, digits = 2)
render_f(mx, digits = list(2, 0, 1))

}

render_sigf(c(123456, 23.2546, 1), 4)


}
