% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-settings.R
\name{set_lattice}
\alias{set_lattice}
\alias{my_theme}
\alias{reset_lattice}
\alias{standard_theme}
\alias{ggplot_theme}
\alias{bw_theme}
\title{Custom Black-and-White Theme for Lattice Plots}
\usage{
set_lattice(theme = bw_theme())

my_theme(
  col = c("#0C5BB0FF", "#15983DFF", "#EC579AFF", "#FA6B09FF", "#667840FF"),
  col.bar = "#708C98FF",
  ...
)

reset_lattice()

standard_theme()

ggplot_theme(
  col = c("#00BA38", "#00BFC4", "#619CFF", "#F564E3", "#F8766D", "#B79F00"),
  col.bar = NULL,
  pch = 15:18,
  lty = 1:3,
  cex = 1,
  cex.main = 1.2,
  cex.axis = 0.8,
  cex.xlab = 1,
  cex.ylab = 1,
  cex.add = 0.8,
  strip.background.col = c("grey80", "grey70", "grey60"),
  ...
)

bw_theme(
  col = grDevices::grey.colors(7, start = 0.3, end = 0.9),
  col.n = 5,
  col.bar = "grey50",
  pch = 15:18,
  lty = 1:3,
  lwd = c(2, rep(1, 6)),
  cex = 1,
  cex.main = 1.2,
  cex.axis = 0.8,
  cex.xlab = 1,
  cex.ylab = 1,
  cex.strip = 0.8,
  cex.add = cex.strip,
  cex.symbol = 0.8,
  strip.lineheight = 1.2,
  strip.background.col = "transparent",
  strip.border.col = "#000000",
  strip.text.col = "#000000",
  bottom = 1,
  top = 1,
  left = 1,
  right = 1,
  box = NULL,
  ...
)
}
\arguments{
\item{theme}{mein lattice theme}

\item{col}{Color palette for plot elements (default: grey scale)}

\item{col.bar}{Color for bars in barcharts}

\item{...}{Additional theme parameters to override}

\item{pch}{Vector of point symbols}

\item{lty}{Vector of line types}

\item{cex, cex.main, cex.axis, cex.strip, cex.symbol, cex.xlab, cex.ylab, cex.add}{Base scaling factor for
title, axis labels, x and y labels, strip labels}

\item{strip.background.col, strip.border.col, strip.text.col}{Background, Border, Text color for strips}

\item{col.n}{Number of colors to generate if col is a palette name}

\item{lwd}{Vector of line widths}

\item{strip.lineheight}{numeric strip lineheight}

\item{bottom, top, left, right}{theme layout.heights and layout.widths Numeric vector of length 4 for plot padding (bottom, left, top, right)}

\item{box}{Logical indicating whether to draw plot borders}
}
\value{
A lattice theme object
}
\description{
Creates a customized black-and-white theme for lattice graphics with sensible defaults
that work well for academic publications and presentations.
}
\examples{

library(lattice)
axis.padding   <- lattice.getOption("axis.padding")
layout.heights <- lattice.getOption("layout.heights")
layout.widths  <- lattice.getOption("layout.widths")

#' Abstand der inneren Labels in effect::plot() (ist ja ein lattice-plot)
lattice.options(axis.padding =list(numeric=0.3))

#' Abstand oben, unten, links, rechts
lattice.options(
  layout.widths =
    list(
      left.padding=list(x=.4),
      right.padding=list(x=1)
    )
  )

#' reset
lattice.options(
  layout.widths = layout.widths,
  layout.heights =layout.heights,
  axis.padding =axis.padding
  )


require(lattice)
#lattice::trellis.par.set(standard_theme())
lattice::trellis.par.set(bw_theme(farbe()))
barchart(yield ~ variety | site, data = barley,
         groups = year, layout = c(1,6), stack = TRUE,
         auto.key = list(space = "right"),
         ylab = "Barley Yield (bushels/acre)",
         scales = list(x = list(rot = 45)))
#'
his <- data.frame(variable = LETTERS[1:5],  value = c(2, 3, 5, 6, 4))

barchart(
  variable ~ value ,
  his,
  xlab = "Prozent nicht zutreffend",
  origin = 0,
  par.settings =
    bw_theme(col.bar  =  "#708C98FF")
)


 #lattice::trellis.par.set(
 #  effects::effectsTheme(
 #
 #    strip.background = list(col = "transparent"),
 #    strip.shingle = list(col = "black"),
 #    clip = list(strip = "off"),
 #    superpose.line = list(lwd = c(2, rep(1, 6))),
 #    col = grey.colors(7, start = 0.3, end = 0.9)
 #    ))


}
