% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_stacked.R, R/tbll_multi.R
\name{prepare_upset_data}
\alias{prepare_upset_data}
\alias{multi_matrix_plot}
\alias{Tbll_multi}
\title{Analyze multi-response items}
\usage{
prepare_upset_data(..., use.level = 1, cutoff = NULL)

multi_matrix_plot(
  data,
  intersect = colnames(data),
  ...,
  ylab.inter = "Intersection Size",
  xlab.set = "Set Size",
  xlab.matrix = ""
)

Tbll_multi(
  ...,
  by = NULL,
  digits = 0,
  include.order = TRUE,
  exclude.last.order = FALSE,
  use.level = 1,
  include.label = TRUE,
  include.n = TRUE,
  include.nr = FALSE,
  include.total = FALSE,
  include.test = FALSE,
  include.normality.tests = FALSE,
  include.custom = NULL,
  include.value = NULL
)
}
\arguments{
\item{...}{Variables to analyze, can be specified as formula, variable names,
or with headings for grouping}

\item{use.level}{Which level to use for analysis (1 for TRUE, 2 for FALSE)}

\item{cutoff}{daten transformieren in prepare_upset_data}

\item{data, intersect, ylab.inter, xlab.set, xlab.matrix}{multi_matrix_plot: Beschriftung in plot}

\item{by}{Grouping variable(s) as formula}

\item{digits}{Number of digits for percentages}

\item{include.order}{Logical, whether to include ordering}

\item{exclude.last.order}{Logical, whether to exclude last order}

\item{include.label}{Logical, whether to include variable labels}

\item{include.n}{Logical, whether to include sample size}

\item{include.nr}{Logical, whether to include number of responses}

\item{include.total}{Logical, whether to include total column}

\item{include.test}{Logical, whether to include statistical tests}

\item{include.normality.tests}{Logical, whether to include normality tests}

\item{include.custom}{Custom content to include}

\item{include.value}{Value to include}
}
\value{
tibble

A tibble with analysis results
}
\description{
Helper for the preparation of `UpSetR::upset` plots.

Stolen from Michal Mrassowski https://github.com/krassowski/complex-upset

Functions for analyzing multi-response items (Likert scales and multiple choice questions).
}
\examples{

 # library(ComplexUpset)
 DF <-
structure(list(
  q2.almdd = structure(
  c(TRUE, FALSE, FALSE, TRUE,TRUE, TRUE, TRUE, TRUE,
  TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE),
                       label = "Alm Dudler"),
  q2.minrl = structure(
  c(TRUE, FALSE,FALSE, FALSE, TRUE, FALSE, FALSE,
  TRUE, TRUE, TRUE, FALSE, FALSE,TRUE, TRUE, TRUE),
                       label= "Mineral"),
  q2.cola = structure(
  c(TRUE,TRUE, TRUE, TRUE, FALSE, TRUE, FALSE,
  FALSE, TRUE, TRUE, TRUE, FALSE, FALSE, TRUE, TRUE),
                      label = "Cola"),
  q2.bier = structure(
  c(TRUE,FALSE, FALSE, TRUE, FALSE, FALSE, TRUE,
  FALSE, FALSE, FALSE,FALSE, TRUE, FALSE, FALSE, FALSE),
                      label = "Bier"),
  q2.wein = structure(
  c(TRUE,FALSE, TRUE, FALSE, FALSE, FALSE, FALSE,
   FALSE, FALSE, FALSE,FALSE, FALSE, FALSE, FALSE, FALSE),
                      label = "Wein")),
  row.names = c(NA,15L),
  class = "data.frame")

dat <- prepare_upset_data(DF,
q2.almdd,q2.minrl,q2.cola,q2.bier,q2.wein )

multi_matrix_plot(dat)

DF <-
  structure(list(
    q2.almdd = structure(
    c(TRUE, FALSE, FALSE, TRUE,TRUE, TRUE, TRUE, TRUE,
    TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE),
                         label = "Alm Dudler"),
    q2.minrl = structure(
    c(TRUE, FALSE,FALSE, FALSE, TRUE, FALSE, FALSE,
    TRUE, TRUE, TRUE, FALSE, FALSE,TRUE, TRUE, TRUE),
                         label= "Mineral"),
    q2.cola = structure(
    c(TRUE,TRUE, TRUE, TRUE, FALSE, TRUE, FALSE, FALSE,
    TRUE, TRUE, TRUE, FALSE, FALSE, TRUE, TRUE),
                        label = "Cola"),
    q2.bier = structure(
    c(TRUE,FALSE, FALSE, TRUE, FALSE, FALSE, TRUE,
    FALSE, FALSE, FALSE,FALSE, TRUE, FALSE, FALSE, FALSE),
                        label = "Bier"),
    q2.wein = structure(
    c(TRUE,FALSE, TRUE, FALSE, FALSE, FALSE, FALSE,
    FALSE, FALSE, FALSE,FALSE, FALSE, FALSE, FALSE, FALSE),
                        label = "Wein")),
    row.names = c(NA,15L),
    class = "data.frame")

DF$sex <- factor(c(1,1,1,1,1,1,1,2,2,2,2,2,2,2,2),
                 1:2, c("f", "m"))

# DF |> aggregat_likert(~ q2.almdd+q2.minrl+q2.cola+q2.bier+q2.wein,
#                       type.wide=TRUE)

DF |> Summarise_multi(~ q2.almdd+q2.minrl+q2.cola+q2.bier+q2.wein,
type.wide=TRUE) |>  
gg_stacked(DF, facet_formula = ~Sex )

DF |> Summarise_multi(~ q2.almdd+q2.minrl+q2.cola+q2.bier+q2.wein,
type.wide=FALSE)

DF |>
  Summarise_multi("Soda (gas)", q2.almdd,
                  q2.minrl,q2.cola, "Alkohol", q2.bier,q2.wein, by=~sex)

typ_name <-DF |>
  Tbll_multi("Soda (gas)", q2.almdd,
             q2.minrl,q2.cola, "Alkohol", q2.bier,q2.wein, by=~sex)

type_formula <- DF |> Tbll_multi(
"Soda (gas)"+q2.almdd + q2.minrl + q2.cola +
 "Alkohol"+q2.bier +q2.wein ~sex)

# type_formula
# typ_name

attr(typ_name, "tbll_likert")

attr(type_formula, "tbll_likert")
}
