% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_xtable.R
\name{Tbll_xtabs_ol}
\alias{Tbll_xtabs_ol}
\alias{Tbll_xtabs_ol.default}
\alias{Tbll_xtabs_ol.glm}
\alias{Tbll_xtabs_ol.table}
\alias{Tbll_xtabs_ol.xtabs}
\title{Kreuztabellen}
\usage{
Tbll_xtabs_ol(x, ...)

\method{Tbll_xtabs_ol}{default}(
  ...,
  include.count = TRUE,
  include.percent = TRUE,
  include.prop.chisq = FALSE,
  include.chisq = FALSE,
  include.fisher = FALSE,
  include.test = any(c(include.fisher, include.chisq, include.prop.chisq)),
  include.correlation = FALSE,
  include.diagnostic = FALSE,
  formula = NULL,
  margin = NULL,
  add.margins = NULL,
  digits = get_opt("percent", "digits"),
  prevalence = NULL,
  addNA = FALSE,
  exclude = if (!addNA) c(NA, NaN),
  drop.unused.levels = FALSE
)

\method{Tbll_xtabs_ol}{glm}(x, thresh = 0.5, ...)

\method{Tbll_xtabs_ol}{table}(...)

\method{Tbll_xtabs_ol}{xtabs}(
  x,
  include.count = TRUE,
  include.percent = TRUE,
  include.prop.chisq = FALSE,
  include.chisq = FALSE,
  include.fisher = FALSE,
  include.test = any(c(include.fisher, include.chisq, include.prop.chisq)),
  include.correlation = FALSE,
  include.diagnostic = FALSE,
  margin = NULL,
  add.margins = NULL,
  digits = get_opt("percent", "digits"),
  prevalence = NULL,
  N_data = sum(x),
  ...
)
}
\arguments{
\item{x}{Objekt  glm, xtab, formula}

\item{...}{alles weitere an prepare.data}

\item{include.count}{Anzahl}

\item{include.percent}{Prozent}

\item{include.prop.chisq, include.chisq, include.fisher, include.test}{Sig. Test}

\item{include.correlation}{Korrelation}

\item{include.diagnostic, prevalence}{Diagnostic}

\item{formula}{an Wide}

\item{margin, add.margins}{Prozent und Total 
add.margins gibt an welche Spalten  geht am addmargins()}

\item{digits}{Nachkommastellen}

\item{addNA, exclude, drop.unused.levels}{an xtabs()'}

\item{lvs, N_data}{internal was fuer eine Tabelle kommt}
}
\value{
list("xtab","fisher_test","diagnostic.test")
}
\description{
Kreuztabellen
}
\examples{

\donttest{

Plank <- data.frame(
  therapy = c("H+A", "H+A+P", "H+A", "H+A+P", "H+A", "H+A+P", "H+A", "H+A+P"),
  mace = c("yes", "yes", "no", "no", "yes", "yes", "no", "no"),
  infarction = c("NSTEMI", "NSTEMI", "NSTEMI", "NSTEMI", "STEMI", "STEMI", "STEMI", "STEMI"),
  freq = c(7, 1, 140, 54, 63, 14, 221, 126)
)
Plank$mace_bin <- ifelse(Plank$mace == "yes", 1, 0)



xtabs(freq ~ therapy + infarction + mace, Plank) |> 
  Tbll_xtabs_ol(digits = 0)
# # OR
# glm_logistic <- glm(mace_bin ~ therapy * infarction,
#                     data=Plank, weights = freq,
#                     family = binomial())
# # andere Methode die Prozent zu berechnen -> glm-logistic
# 
# # RR
#  glm_poisson <- glm(freq ~ therapy * mace + infarction * mace,data = Plank,
#                     family = poisson())
#  Tbll_effect(glm_logistic, digits = 3, include.ci = FALSE)
#  Tbll_effect(glm_poisson, digits = 3, include.ci = FALSE)

set_opt(percent=list(digits=0))


data(infert, package = "datasets")
infert2<- infert
infert2$case  <- factor(infert2$case ,1:0, c("case", "control") )
infert2$spontaneous <- factor(infert2$spontaneous)
infert2$induced2    <- factor(infert2$induced==0)

tab_1 <- xtabs( ~  case, infert2)
tab_2x2 <- xtabs( ~ induced2 + case, infert2)
tab_3x2 <- xtabs( ~ induced + case, infert2)
tab_3x3 <- xtabs( ~ induced + education, infert2)
tab_3x3x2 <- xtabs( ~ induced + education + case, infert2)



#Tbll_xtabs_ol(tab_1)

Tbll_xtabs_ol(tab_2x2)
Tbll_xtabs_ol(tab_3x2)
Tbll_xtabs_ol(tab_3x3)
Tbll_xtabs_ol(tab_3x3x2)

Tbll_xtabs_ol(  induced + education ~ case, infert2)
Tbll_xtabs_ol(  induced + education ~ case, infert2, include.test = TRUE)

Tbll_xtabs_ol_diagnostic(tab_2x2)

# tab <- matrix(c(94, 40, 39, 40), ncol = 2, byrow = TRUE)
# tbll_extract(caret::confusionMatrix(tab))
# #tbll_extract(epiR::epi.tests(tab) )

Tbll_xtabs_ol(  induced  ~ case, infert2 )
Tbll_xtabs_ol(  induced  ~ case, infert2, margin= "case" )
Tbll_xtabs_ol(  induced  ~ case, infert2, margin= "case", add.margins="induced")
}

}
