% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{to_formula}
\alias{to_formula}
\title{Convert to formula
to_formula(letters[1:4], NULL)
~a + b + c + d
to_formula(letters[1:4], letters[5:6])
a + b + c + d ~ e + f}
\usage{
to_formula(measure_vars, group_vars)
}
\arguments{
\item{measure_vars}{Measurement variables for LHS}

\item{group_vars}{Grouping variables for RHS}
}
\description{
Convert to formula
to_formula(letters[1:4], NULL)
~a + b + c + d
to_formula(letters[1:4], letters[5:6])
a + b + c + d ~ e + f
}
\keyword{internal}
