% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{which_digits}
\alias{which_digits}
\title{Determine appropriate rounding digits}
\usage{
which_digits(extras, measure, names)
}
\arguments{
\item{extras}{Character vector of extra specifications}

\item{measure}{Vector of statistical measures}

\item{names}{Character vector of variable names}
}
\description{
Determine appropriate rounding digits
}
\examples{
# example code

stp25tools2:::which_digits("mean,5,anova", "factor", "BMI")
#c(BMI = 5)
stp25tools2:::which_digits("mean,anova", "numeric", "BMI")
#c(BMI = 2)

#' @noRd
}
\keyword{internal}
