% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{which_output}
\alias{which_output}
\title{which_output}
\usage{
which_output()
}
\value{
values are text, latex, html, markdown_html markdown, pandoc, and rst
}
\description{
which_output
}
\examples{

which_output()

}
