% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R
\name{APA}
\alias{APA}
\alias{APA.default}
\alias{APA.lm}
\alias{APA.glm}
\alias{APA.htest}
\alias{APA.xtabs}
\alias{APA.table}
\alias{APA.summary.table}
\alias{APA.psych}
\alias{APA.psych_alpha}
\alias{APA.coxph}
\alias{APA.survdiff}
\alias{APA.QuadTypeIndependenceTest}
\alias{APA.ScalarIndependenceTest}
\alias{APA.numeric}
\alias{Mean}
\alias{Median}
\alias{Count}
\alias{Percent}
\title{Text-Ausgabe von Tests}
\usage{
APA(x, ...)

\method{APA}{default}(x, ...)

\method{APA}{lm}(x, include.r = TRUE, ...)

\method{APA}{glm}(x, ...)

\method{APA}{htest}(x, ...)

\method{APA}{xtabs}(x, ...)

\method{APA}{table}(x, ...)

\method{APA}{summary.table}(x, ...)

\method{APA}{psych}(x, ...)

\method{APA}{psych_alpha}(x, ...)

\method{APA}{coxph}(x, ...)

\method{APA}{summary.table}(x, ...)

\method{APA}{survdiff}(x, ...)

\method{APA}{QuadTypeIndependenceTest}(x, ...)

\method{APA}{ScalarIndependenceTest}(x, ...)

\method{APA}{numeric}(x, ...)

Mean(x, digits = get_opt("mean", "digits"))

Median(x, digits = get_opt("median", "digits"))

Count(x, sep = ", and ", ...)

Percent(x, style = get_opt("percent", "style"), sep = ", and ", ...)
}
\arguments{
\item{x}{Objekt fit, formula usw}

\item{...}{digits, data, usw}

\item{include.r}{APA.lm: R-Squar}
}
\value{
Character Vector mit einem oder meheren Eintraegen
}
\description{
Ausgabe von APA-Style formatiertem Text.
}
\examples{


APA(glm(vs ~ mpg, mtcars, family = binomial()))
APA(lm(mpg ~ drat + wt + qsec, mtcars))
APA(aov(mpg ~ drat + wt + qsec, mtcars))
APA(wilcox.test(mpg ~ vs, mtcars))
APA(t.test(mpg ~ vs, mtcars))

a <- letters[1:3]
APA(summary(table(a, sample(a))))

TeaTasting <-
  matrix(c(3, 1, 1, 3),
         nrow = 2,
         dimnames = list(
           Guess = c("Milk", "Tea"),
           Truth = c("Milk", "Tea")
         ))
APA(fisher.test(TeaTasting, alternative = "greater"))

# Test Survival Curve Differences
# require(survival)
# APA(survdiff(Surv(futime, fustat) ~ ecog.ps, data = ovarian))



# require(coin)
# davis <- matrix(c(3, 6, 2, 19),
#               nrow = 2, byrow = TRUE)
# davis <- as.table(davis)
## Asymptotic Pearson chi-squared test
# diffusion <- data.frame(
#  pd = c(
#    0.80,    0.83,    1.89,    1.04,    1.45,
#   1.38,    1.91,    1.64,    0.73,    1.46,
#    1.15,    0.88,    0.90,    0.74,    1.21),
#  age = factor(rep(c( "At term", "12-26 Weeks"), c(10, 5)))
# )
#
## Exact Wilcoxon-Mann-Whitney test
## Hollander and Wolfe (1999, p. 111)
## (At term - 12-26 Weeks)
# wt <- wilcox_test(pd ~ age,
#                 data = diffusion,
#                 distribution = "exact",
#                 conf.int = TRUE)
#
# APA(xt <- chisq_test(davis))
# APA(wt)

g <- gl(3, 8, labels = c("Omnivor", "Vegetarian", "Vegan"))

x <- 1:20
Mean(x)
Median(x)

Count(g)
Percent(g)
}
