% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{Praevalenz}
\alias{Praevalenz}
\title{Praevalenz}
\usage{
Praevalenz(x, by, ...)
}
\arguments{
\item{x}{logical}

\item{by}{not used}

\item{...}{not used}
}
\value{
vector
}
\description{
Clopper Pearson method for confidence intervals
}
\examples{

DF<- data.frame(
  sex = gl(2, 50, labels = c("male", "female")),
  Adipositas = rbinom(n=100, 1, prob=.2),
  Bewegungsmangel =rbinom(n=100, 1, prob=.4),
  Nikotinabusus = rbinom(n=100, 1, prob=.3)
)

DF |> Summarise(
  Adipositas,
  Bewegungsmangel,
  Nikotinabusus,
  #  by=~sex,
  key = "Risikofaktor",
  fun = Praevalenz,
  # formula= Risikofaktor ~ sex
)

}
