% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll-test.R
\name{Tbll_test}
\alias{Tbll_test}
\alias{auto_test}
\alias{conTest}
\alias{catTest}
\alias{ordTest}
\alias{spearmanTest2}
\alias{WilkoxTest2}
\alias{KruskalTest2}
\alias{Aov2}
\alias{TTest2}
\alias{chisqTest2}
\alias{gml_binomial}
\alias{fisherTest2}
\title{Tbll_test}
\usage{
Tbll_test(...)

auto_test(
  x = 1:16,
  by = gl(2, 8, labels = c("m", "f")),
  measure = NULL,
  measure.test = "contest"
)

conTest(x, data, test_name = TRUE)

catTest(x, data, include.test = "chisq")

ordTest(x, data)

spearmanTest2(x, data)

WilkoxTest2(x, data)

KruskalTest2(x, data)

Aov2(x, data)

TTest2(x, data)

chisqTest2(x, data)

gml_binomial(x, data)

fisherTest2(x, data)
}
\arguments{
\item{...}{formula und daten}

\item{x}{vector , formula x~gruppe}

\item{by}{vector (group)}

\item{measure}{not used}

\item{measure.test}{test statistics}

\item{data}{data.frame}

\item{test_name}{c("contest", "wilcox","utest",
"htest", "kruskal", "ttest","aov", "anova")}

\item{include.test}{Tbll_desc kann entweder logical oder den Test uebergeben}
}
\value{
data.frame

formatierter String, matrix with one row
}
\description{
auto_test(): Interne Funktion fuer Sig-Test in  Tbll_desc.  include.custom = auto_test

conTest(): spearmanTest2, WilkoxTest2,
KruskalTest2,TTest2, Aov2

catTest(): chisqTest2, fisherTest2, gml_binomial

ordTest(). Logistic Regression Model  >Hier gibt es noch einen Fehler!!!

spearmanTest2(): Uses midranks in case of ties, as described by Hollander and Wolfe.
P-values for Spearman, Wilcoxon, or Kruskal-Wallis tests are
approximated by using the t or F distributions.

WilkoxTest2(): wilcox.test

KruskalTest2(): kruskal.test

Aov2():  car::Anova(lm(...))

TTest2(): t.test

chisqTest2(): chisq.test

gml_binomial(): glm

fisherTest2(): fisher.test
}
\examples{

Tbll_test(breaks + tension ~ wool , data = warpbreaks)



 dat <- data.frame(
   m1 = c(1, 2, 1, 3, 1, 4, 1, 1,
       3, 5, 3, 1, 3, 6, 3, 1),
   geschl = gl(2, 8, labels = c("m", "f"))
 )
 dat$m2 <- cut(dat$m1, 2)


 dat |> Tbll_desc(
  m1[median, 1],
  m2,
  by = ~ geschl,
  include.custom = auto_test,
  # include.test = TRUE,
  include.total = TRUE
 )

\donttest{

dat <- data.frame(
m1 = c(1, 2, 1, 3, 1, 4, 1, 1,
       3, 5, 3, 1, 3, 6, 3, 1),
geschl = gl(2, 8, labels = c("m", "f"))
)
dat$m2 <- cut(dat$m1, 2)
#'
stp25tools2:::conTest(m1  ~ geschl, dat)
stp25tools2:::catTest(m1  ~ geschl, dat)
#stp25tools2:::ordTest(m1  ~ geschl, dat) Fehler???
stp25tools2:::spearmanTest2(m1  ~ geschl, dat)
stp25tools2:::TTest2(m1  ~ geschl, dat)
stp25tools2:::Aov2(m1  ~ geschl, dat)
stp25tools2:::KruskalTest2(m1  ~ geschl, dat)
stp25tools2:::WilkoxTest2(m1  ~ geschl, dat)
stp25tools2:::chisqTest2(m1  ~ geschl, dat)
stp25tools2:::chisqTest2( ~ m1 + geschl, dat)
stp25tools2:::fisherTest2( ~ m1 + geschl, dat)
stp25tools2:::fisherTest2( ~ m2 + geschl, dat)

# Wilcoxon.Test       "F(1, 14)=3.20, p=.095"
# V2                  "Error wrong formula!"
# Logistic.Regression "LR(1)=3.14, p=.077"
# Wilcoxon.Test.1     "F(1, 14)=3.20, p=.095"
# T.Tests             "T(12)=-1.87, p=.086"
# ANOVA               "F(1, 14)=3.49, p=.083"
# Kruskal.Wallis.Test "H(1)=2.79, p=.095"
# Wilcoxon.Test.2     "U=17.00, p=.106"
# V9                  "Error wrong formula!"
# Pearson.Chi.squared "X2(5)=7.09, p=.214"
# V11                 "wrong dim for fisher-test"
# Fisher.Exact.Test   "OR=2.21, p=1.000"

}
}
